% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdf.R
\name{checkdf}
\alias{checkdf}
\title{Checks dataframe dimensions and column names.}
\usage{
checkdf(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
The dimensions of \code{x} and column names of \code{x}
}
\description{
Checks dataframe dimensions and column names.
}
\examples{
checkdf(mtcars)

}
