% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_raw.R
\docType{data}
\name{hospital_raw}
\alias{hospital_raw}
\title{hospital_raw}
\format{A data frame with 5344 rows and 29 variables}
\source{
\url{https://data.medicare.gov/Hospital-Compare/Hospital-General-Information/xubh-q36u}
}
\usage{
hospital_raw
}
\description{
Raw dataset about hospitals across the U.S.
For more information, please go to the source url.
}
\keyword{datasets}
