% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection.R
\docType{data}
\name{infection}
\alias{infection}
\title{infection}
\format{A clean data frame with 3119 rows and 2 variables:
\describe{
  \item{facility_id}{6 digit unique reference code for each hospital}
  \item{infection_score}{Average score on in-hospital infections compared with the national average. 1 is below national average, 2 is not different from national average, 3 is better than national average.}
  ...
}}
\source{
\url{https://data.medicare.gov/Hospital-Compare/Healthcare-Associated-Infections-Hospital/77hc-ibv8}
}
\usage{
infection
}
\description{
Dataset about 3119 different hospitals in the U.S. on in-hospital infections.
}
\keyword{datasets}
