% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_raw.R
\docType{data}
\name{infection_raw}
\alias{infection_raw}
\title{infection_raw}
\format{A data frame with 171252 rows and 16 variables about different hospital related infections}
\source{
\url{https://data.medicare.gov/Hospital-Compare/Healthcare-Associated-Infections-Hospital/77hc-ibv8}
}
\usage{
infection_raw
}
\description{
Dataset of 171252 observations of 16 variables.
For more information, please check the source url.
}
\keyword{datasets}
