% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_q.R
\docType{data}
\name{survey_q}
\alias{survey_q}
\title{survey_q}
\format{A clean data frame with 22 rows and 2 variables:
\describe{
  \item{code}{Reference code for questions}
  \item{description}{Question description}
  ...
}}
\source{
\url{https://data.medicare.gov/Hospital-Compare/Patient-survey-HCAHPS-Hospital/dgck-syfz}
}
\usage{
survey_q
}
\description{
Clean dataset about survey questions.
Source url has information about the raw dataframe used to create this dataset
}
\keyword{datasets}
