% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_raw.R
\docType{data}
\name{survey_raw}
\alias{survey_raw}
\title{survey_raw}
\format{A data frame with 455142 rows and 23 variables}
\source{
\url{https://data.medicare.gov/Hospital-Compare/Patient-survey-HCAHPS-Hospital/dgck-syfz}
}
\usage{
survey_raw
}
\description{
Dataset about hospital survey in raw format with 455142 observations of 23 variables.
For more information, please check the source url.
}
\keyword{datasets}
