% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquevalue.R
\name{uniquevalue}
\alias{uniquevalue}
\title{Returns the number of unique values in a dataframe column
Generates a dataframe with column names and the number of unique values in each column.}
\usage{
uniquevalue(x)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A dataframe with two columns. First column has the variables in input dataframe. Second column has the number of unique values in each variable of the input dataframe.
}
\description{
Returns the number of unique values in a dataframe column
Generates a dataframe with column names and the number of unique values in each column.
}
\examples{
df <- data.frame(a = rnorm(10), b = c(1:5, 1:5))
uniquevalue(df)

}
